Call tracing-subscriber only from directory binary
